Module binjs_io::entropy

An encoding using entropy coding. A format based on Entropy coding

We generate a table of probabilities that let us determine the probability of a given node appearing at a given position in the AST. We then use a mechanism of entropy coding to represent the actual AST as a combination of these probabilities.

See module format for details on the binary format.

Modules

dictionary

format
About this file

probabilities

read
Reading a compressed stream. An entropy decoder

rw
Tools shared by read and write. Constants and data structures shared between entropy read and entropy write.

write
Writing to a compressed stream. An entropy encoder.

Structs

FormatProvider

Options