Module binjs_io::entropy::dictionary

Re-exports

pub use entropy::predict::Instances;

Structs

Dictionary

DictionaryBuilder
A structure used to build a dictionary based on a sample of files.

DictionaryFamily
A family of dictionaries.

FilesContaining
A newtype for usize used to count the number of files containing some item.

InstancesInFile
A newtype for usize used to count the number of some item in a given file.

LinearTable
A data structure designed to cache information accessible either by index or by value.

Options
Options for creating a dictionary.

UserExtensibleData
A container for all the user-extensible data in files, e.g. string literals or numbers.

Enums

Fetch
Result of a fetch operation in a LinearTable.

TableRef
An index in an LinearTable.

Type Definitions

IOPath