Struct binjs_io::entropy::rw::PreludeStreams

pub struct PreludeStreams<T> {
    pub identifier_names: T,
    pub identifier_names_len: T,
    pub property_keys: T,
    pub property_keys_len: T,
    pub string_literals: T,
    pub string_literals_len: T,
    pub interface_names: T,
    pub string_enums: T,
    pub list_lengths: T,
    pub floats: T,
    pub unsigned_longs: T,
}

Prelude data.

The prelude contains streams to read or write dictionary extensions for user-extensible symbols (aka "prelude dictionaries").

Examples of T: LazyStream (for writing) or Cursor<Vec<u8>> (for reading).

Fields
identifier_names: T

Instances of IdentifierName.

To aid with compression, we typically store user-extensible strings as two distinct data structures:

a single string, containing all concatenated strings without delimiter;

a list of lengths, used to split the string into the successive strings.

identifier_names_len: Tproperty_keys: T

Instances of PropertyKey

To aid with compression, we typically store user-extensible strings as two distinct data structures:

a single string, containing all concatenated strings without delimiter;

a list of lengths, used to split the string into the successive strings.

property_keys_len: Tstring_literals: T

Instances of string literals.

To aid with compression, we typically store user-extensible strings as two distinct data structures:

a single string, containing all concatenated strings without delimiter;

a list of lengths, used to split the string into the successive strings.

string_literals_len: Tinterface_names: T

Instances of InterfaceName

string_enums: T

Instances of string enums.

list_lengths: T

Instances of list lengths.

floats: T

Instances of floating-point numbers.

unsigned_longs: T

Instances of unsigned longs.

Methods

impl<T> PreludeStreams<T>

pub fn with<F>(f: F) -> Self where
F: Fn(&str) -> T,

Create a new PreludeStreams.

pub fn into_iter(self) -> impl Iterator<Item = (&'static str, T)>

Iterate through fields of PreludeStreams.

pub fn get_mut_b(&mut self, field_name: &[u8]) -> Option<&mut T>

Access a field by its name, specified as a sequence of bytes.

This method is typically used to simplify parsing a file that contains sections explicitly labelled "identifier_names", "property_keys", etc. In such case, field_name is expected to be a user input.

Return None if field_name is not one of the field names.