Enum binjs_io::Format

pub enum Format {
    Simple,
    Multipart {
        targets: Targets,
        stats: Rc<RefCell<Statistics>>,
    },
    XML,
    JSON,
    Entropy {
        options: Options,
    },
}

All the formats available for encoding/decoding.

Variants

Simple

Multipart

Fields of Multipart

targets: Targetsstats: Rc<RefCell<Statistics>>

XML

JSON

Entropy

Fields of Entropy

options: Options

Methods

impl Format

pub fn simple() -> Self

pub fn randomize_options<R: Rng>(self, rng: &mut R) -> Self

Pick a random set of options.

Used for testing.

pub fn name(&self) -> String

Return a human-readable name for this format.

pub fn with_sections<F, E>(&mut self, f: F) -> Result<(), E> where
F: FnMut(&mut CompressionTarget, &str) -> Result<(), E>,

pub fn subcommand<'a, 'b>() -> App<'a, 'b>

Returns command-line argument for advanced. FormatProvider's subcommands are hidden behind "advanced" command.

pub fn from_matches(spec: &Spec, matches: &ArgMatches) -> Result<Self, Error>

Create a Format based on command-line arguments.

Pick the first format provider that was invoked by matches as a subcommand. If none, pick the default provider, without any command-line arguments.