Module binjs_io::io

Definition of TokenReader/TokenWriter traits. Mid-level API used to write individual components of an AST tree to a stream of tokens.

Simple implementations may write bytes as they come, while complex ones may decide to provide sophisticated compression.

In practice, this API is kept as a trait to simplify unit testing and experimentation of sophisticated compression schemes.

Modules

statistics
Utilities to collect statistics about the data written.

Structs

TokenWriterTreeAdapter

Traits

Deserialization

FileStructurePrinter
An API for printing the binary representation and its structural interpretation of the file.

InnerDeserialization

RootedTokenSerializer

Serialization

TokenReader
An API for reading tokens.

TokenSerializer

TokenSerializerFamily

TokenWriter
Build an in-memory representation of a BinTree.

TokenWriterWithTree

Type Definitions

Path