Trait binjs_io::io::TokenReader

pub trait TokenReader: FileStructurePrinter where
    Self::ListGuard: Guard,
    Self::TaggedGuard: Guard,
    Self::UntaggedGuard: Guard,
    Self: Sized, {
    type ListGuard;
    type TaggedGuard;
    type UntaggedGuard;
    fn string_at(
        &mut self, 
        _path: &Path
    ) -> Result<Option<SharedString>, TokenReaderError>;

    fn float_at(
        &mut self, 
        _path: &Path
    ) -> Result<Option<f64>, TokenReaderError>;

    fn unsigned_long_at(
        &mut self, 
        _path: &Path
    ) -> Result<u32, TokenReaderError>;

    fn bool_at(
        &mut self, 
        _path: &Path
    ) -> Result<Option<bool>, TokenReaderError>;

    fn offset_at(&mut self, _path: &Path) -> Result<u32, TokenReaderError>;

    fn list_at(
        &mut self, 
        _path: &Path
    ) -> Result<(u32, Self::ListGuard), TokenReaderError>;

    fn tagged_tuple_at(
        &mut self, 
        _path: &Path
    ) -> Result<(InterfaceName, Option<Rc<Box<[FieldName]>>>, Self::TaggedGuard), TokenReaderError>;

    fn untagged_tuple_at(
        &mut self, 
        _path: &Path
    ) -> Result<Self::UntaggedGuard, TokenReaderError>;

    fn poison(&mut self) { ... }

    fn string_enum_at(
        &mut self, 
        path: &Path
    ) -> Result<SharedString, TokenReaderError> { ... }

    fn identifier_name_at(
        &mut self, 
        path: &Path
    ) -> Result<Option<IdentifierName>, TokenReaderError> { ... }

    fn property_key_at(
        &mut self, 
        path: &Path
    ) -> Result<Option<PropertyKey>, TokenReaderError> { ... }
}

An API for reading tokens.

Note that a TokenReader by itself cannot determine the nature of the following token. Rather, the driver of the TokenReader must be able to deduce the nature of the following token from what it has previously read.

All the reading methods offer a version suffixed with _at(path: &Path), which lets the reader determine what item we're reading in the AST. This may be used both for debugging purposes and for encodings that depend on the current position in the AST (e.g. entropy coding).

Associated Types

type ListGuard

A guard, used to make sure that the consumer has properly read a list.

See the documentation of self.list.

type TaggedGuard

A guard, used to make sure that the consumer has properly read a tagged tuple.

See the documentation of self.tagged_tuple.

type UntaggedGuard

A guard, used to make sure that the consumer has properly read an untagged tuple.

See the documentation of self.untagged_tuple.

Required Methods

fn string_at(
&mut self,
_path: &Path
) -> Result<Option<SharedString>, TokenReaderError>

Read a single UTF-8 string.

The returned string MUST be valid UTF-8.

fn float_at(&mut self, _path: &Path) -> Result<Option<f64>, TokenReaderError>

Read a single f64. Note that all user-level numbers are f64.

fn unsigned_long_at(&mut self, _path: &Path) -> Result<u32, TokenReaderError>

Read a single u32.

fn bool_at(&mut self, _path: &Path) -> Result<Option<bool>, TokenReaderError>

Read a single bool.

fn offset_at(&mut self, _path: &Path) -> Result<u32, TokenReaderError>

Read a single number of bytes.

fn list_at(
&mut self,
_path: &Path
) -> Result<(u32, Self::ListGuard), TokenReaderError>

Start reading a list.

Returns a guard for that list and the number of elements in the list. Once the list is read entirely, callers MUST call guard.done() to ensure that the list was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.

fn tagged_tuple_at(
&mut self,
_path: &Path
) -> Result<(InterfaceName, Option<Rc<Box<[FieldName]>>>, Self::TaggedGuard), TokenReaderError>

Start reading a tagged tuple. If the stream was encoded properly, the tag is attached to an ordered tuple of fields that may be extracted in order.

Returns the tag name, the ordered array of fields in which the contents must be read, and a guard for that tuple. Once the tuple is read entirely, callers MUST call guard.done() to ensure that the tuple was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.

fn untagged_tuple_at(
&mut self,
_path: &Path
) -> Result<Self::UntaggedGuard, TokenReaderError>

Start reading an untagged tuple.

Once the tuple is read entirely, callers MUST call guard.done() to ensure that the tuple was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.

Provided Methods

fn poison(&mut self)

Poison the reader, ensuring that it will never be used for reading again.

fn string_enum_at(
&mut self,
path: &Path
) -> Result<SharedString, TokenReaderError>

Read a single UTF-8 value from a string enumeration.

The default implementation uses `self.string``, but some encodings may use the extra information e.g. to represent the enumeration by an index in the list of possible values, or to encode string enums as interfaces.

The returned string MUST be valid UTF-8.

fn identifier_name_at(
&mut self,
path: &Path
) -> Result<Option<IdentifierName>, TokenReaderError>

Read a single identifier name.

The default implementation uses self.string_at, but some encodings may use the extra information e.g. to represent the identifier as a DeBruijn index.

fn property_key_at(
&mut self,
path: &Path
) -> Result<Option<PropertyKey>, TokenReaderError>

Read a single property name.

The default implementation uses self.string_at, but some encodings may use the extra information e.g. to initialize dictionaries.