Trait binjs_io::io::TokenWriter

pub trait TokenWriter {
    type Data: AsRef<[u8]>;
    fn done(self) -> Result<Self::Data, TokenWriterError>;

    fn enter_tagged_tuple_at(
        &mut self, 
        _tag: &InterfaceName, 
        _children: &[&FieldName], 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn enter_list_at(
        &mut self, 
        _len: usize, 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn string_at(
        &mut self, 
        value: Option<&SharedString>, 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn string_enum_at(
        &mut self, 
        value: &SharedString, 
        path: &Path
    ) -> Result<(), TokenWriterError>;

    fn float_at(
        &mut self, 
        value: Option<f64>, 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn unsigned_long_at(
        &mut self, 
        value: u32, 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn bool_at(
        &mut self, 
        value: Option<bool>, 
        _path: &Path
    ) -> Result<(), TokenWriterError>;

    fn offset_at(&mut self, _path: &Path) -> Result<(), TokenWriterError>;

    fn exit_tagged_tuple_at(
        &mut self, 
        _tag: &InterfaceName, 
        _children: &[&FieldName], 
        _path: &Path
    ) -> Result<(), TokenWriterError> { ... }

    fn exit_list_at(&mut self, _path: &Path) -> Result<(), TokenWriterError> { ... }

    fn property_key_at(
        &mut self, 
        value: Option<&PropertyKey>, 
        path: &Path
    ) -> Result<(), TokenWriterError> { ... }

    fn identifier_name_at(
        &mut self, 
        value: Option<&IdentifierName>, 
        path: &Path
    ) -> Result<(), TokenWriterError> { ... }
}

Build an in-memory representation of a BinTree.

Implementations may for instance introduce atoms, maximal sharing, etc.

All the reading methods offer a version suffixed with _at(..., path: &Path), which lets the writer determine what item we're reading in the AST. This may be used both for debugging purposes and for encodings that depend on the current position in the AST (e.g. entropy coding).

Associated Types

type Data: AsRef<[u8]>

The type of data generated by this writer. Typically some variant of Vec<u8>.

Required Methods

fn done(self) -> Result<Self::Data, TokenWriterError>

Finish writing, produce data.

fn enter_tagged_tuple_at(
&mut self,
_tag: &InterfaceName,
_children: &[&FieldName],
_path: &Path
) -> Result<(), TokenWriterError>

Write a tagged tuple.

The number of items is specified by the grammar, so it MAY not be recorded by the TokenWriter.

By convention, a null tagged tuple is the special tagged tuple "null", with no children.

fn enter_list_at(
&mut self,
_len: usize,
_path: &Path
) -> Result<(), TokenWriterError>

Write a list.

By opposition to a tuple, the number of items is variable and MUST be somehow recorded by the TokenWriter.

fn string_at(
&mut self,
value: Option<&SharedString>,
_path: &Path
) -> Result<(), TokenWriterError>

Write a single UTF-8 string.

If specified, the string MUST be UTF-8.

fn string_enum_at(
&mut self,
value: &SharedString,
path: &Path
) -> Result<(), TokenWriterError>

Write a single UTF-8 value from a string enumeration.

The default implementation uses `self.string``, but some encodings may use the extra information e.g. to represent the enumeration by an index in the list of possible values, or to encode string enums as interfaces.

fn float_at(
&mut self,
value: Option<f64>,
_path: &Path
) -> Result<(), TokenWriterError>

Write a single number.

fn unsigned_long_at(
&mut self,
value: u32,
_path: &Path
) -> Result<(), TokenWriterError>

Write a single u32.

fn bool_at(
&mut self,
value: Option<bool>,
_path: &Path
) -> Result<(), TokenWriterError>

Write single bool.

fn offset_at(&mut self, _path: &Path) -> Result<(), TokenWriterError>

Write the number of bytes left in this tuple.

Provided Methods

fn exit_tagged_tuple_at(
&mut self,
_tag: &InterfaceName,
_children: &[&FieldName],
_path: &Path
) -> Result<(), TokenWriterError>

fn exit_list_at(&mut self, _path: &Path) -> Result<(), TokenWriterError>

fn property_key_at(
&mut self,
value: Option<&PropertyKey>,
path: &Path
) -> Result<(), TokenWriterError>

fn identifier_name_at(
&mut self,
value: Option<&IdentifierName>,
path: &Path
) -> Result<(), TokenWriterError>