Struct binjs_io::TrivialGuard

pub struct TrivialGuard<Error> {
    pub finalized: bool,
    // some fields omitted
}

Trivial implementation of a guard.

This implementation serves as a placeholder or as a building block for more sophisticated implementations: it does not check anything meaningful in done() but ensures that done() is eventually called.

Fields
finalized: bool

true once done() has been called, false otherwise.

Methods

impl<E> TrivialGuard<E>

pub fn new() -> Self

Create a TrivialGuard.

If the TrivialGuard is dropped before done() is called or self.finalized is set to true, the drop will cause an assertion failure.