Trait binjs_io::TokenWriter

pub trait TokenWriter where
    Self::Error: Debug,
    Self::Statistics: Display + Sized + Add + Default, {
    type Tree;
    type Statistics;
    type Data: AsRef<[u8]>;
    type Error;
    fn done(self) -> Result<(Self::Data, Self::Statistics), Self::Error>;

    fn tagged_tuple(
        &mut self, 
        tag: &str, 
        _: &[(&str, Self::Tree)]
    ) -> Result<Self::Tree, Self::Error>;

    fn untagged_tuple(
        &mut self, 
        _: &[Self::Tree]
    ) -> Result<Self::Tree, Self::Error>;

    fn list(&mut self, _: Vec<Self::Tree>) -> Result<Self::Tree, Self::Error>;

    fn string(&mut self, _: Option<&str>) -> Result<Self::Tree, Self::Error>;

    fn float(&mut self, _: Option<f64>) -> Result<Self::Tree, Self::Error>;

    fn bool(&mut self, _: Option<bool>) -> Result<Self::Tree, Self::Error>;

    fn offset(&mut self) -> Result<Self::Tree, Self::Error>;
}

Build an in-memory representation of a BinTree.

Implementations may for instance introduce atoms, maximal sharing, etc.

Associated Types

type Tree

The type of trees manipulated by this writer.

type Statistics

Statistics produced by this writer.

type Data: AsRef<[u8]>

The type of data generated by this writer. Typically some variant of Vec<u8>.

type Error

An error returned by this writer.

Note that errors are not recoverable within the life of this TokenWriter.

Required Methods

fn done(self) -> Result<(Self::Data, Self::Statistics), Self::Error>

Finish writing, produce data.

fn tagged_tuple(
&mut self,
tag: &str,
_: &[(&str, Self::Tree)]
) -> Result<Self::Tree, Self::Error>

Write a tagged tuple.

The number of items is specified by the grammar, so it MAY not be recorded by the TokenWriter.

The interface MUST have a Tag.

fn untagged_tuple(
&mut self,
_: &[Self::Tree]
) -> Result<Self::Tree, Self::Error>

Write an untagged tuple.

The number of items is specified by the grammar, so it MAY not be recorded by the TokenWriter.

fn list(&mut self, _: Vec<Self::Tree>) -> Result<Self::Tree, Self::Error>

Write a list.

By opposition to a tuple, the number of items is variable and MUST be somehow recorded by the TokenWriter.

fn string(&mut self, _: Option<&str>) -> Result<Self::Tree, Self::Error>

Write a single UTF-8 string.

If specified, the string MUST be UTF-8.

fn float(&mut self, _: Option<f64>) -> Result<Self::Tree, Self::Error>

Write a single number.

fn bool(&mut self, _: Option<bool>) -> Result<Self::Tree, Self::Error>

Write single bool.

fn offset(&mut self) -> Result<Self::Tree, Self::Error>

Write the number of bytes left in this tuple.